customer-data-analytics/api/requirements.txt
# API Dependencies
# FastAPI and related
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Configuration and utilities
python-dotenv>=1.0.0
loguru>=0.7.0
redis>=5.0.0

# CORS and middleware
fastapi-cors>=0.0.6

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# ML libraries
xgboost>=3.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# ML Engine integration (installed separately)
